+++ examples/runner-default/versions.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = "4.63.0"
+      version = "4.64.0"
     }
     local = {
       source  = "hashicorp/local"